refactor: Separar métodos de request para `RedeHttpClient`
